Refactor/#210: auth 관련 로직 api/query/model 정리
