+++ src/apis/user/user.queries.ts
@@ -1,17 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { useNavigate } from 'react-router-dom';
 
-import {
-  postNicknameCheck,
-  deleteAccount,
-  getBoardList,
-  getFavoriteBoardList,
-  getToolList,
-  getUserInfo,
-  logout,
-  patchInfo,
-} from './user.api';
-import { logout as handleLogout } from '@apis/index';
+import { postNicknameCheck, getBoardList, getFavoriteBoardList, getToolList, getUserInfo, patchInfo } from './user.api';
 
 export const MYPAGE_QUERY_KEY = {
   MY_INFO: (userId: number) => ['myInfo', userId], // 개인정보
@@ -102,57 +91,6 @@ export const useFavoriteToolQuery = () => {
     staleTime: 1000 * 60 * 10,
     gcTime: 1000 * 60 * 60 * 24,
     enabled: !!userId,
-  });
-};
-
-// 회원 탈퇴
-export const useAccountDeleteMutation = () => {
-  const userItem = localStorage.getItem('user');
-  const userData = userItem ? JSON.parse(userItem) : null;
-  const userId = userData?.accessToken || null;
-
-  const queryClient = useQueryClient();
-  const navigate = useNavigate();
-
-  return useMutation({
-    mutationFn: () => deleteAccount(),
-    onSuccess: () => {
-      if (userId) {
-        // userId와 관련된 모든 쿼리 무효화
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_INFO(userId) });
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_POST_LIST(userId) });
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_FAVORITE_POST_LIST(userId) });
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_FAVORITE_TOOL_LIST(userId) });
-      }
-
-      // localStorage에서 'user' 삭제
-      localStorage.removeItem('user');
-      navigate('/');
-    },
-  });
-};
-
-// 로그아웃
-export const useLogoutMutation = () => {
-  const userItem = localStorage.getItem('user');
-  const userData = userItem ? JSON.parse(userItem) : null;
-  const userId = userData?.accessToken || null;
-
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: () => logout(),
-    onSuccess: () => {
-      if (userId) {
-        // userId와 관련된 모든 쿼리 무효화
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_INFO(userId) });
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_POST_LIST(userId) });
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_FAVORITE_POST_LIST(userId) });
-        queryClient.invalidateQueries({ queryKey: MYPAGE_QUERY_KEY.MY_FAVORITE_TOOL_LIST(userId) });
-      }
-      queryClient.clear();
-      handleLogout();
-    },
   });
 };
