+++ src/pages/communityWrite/CommunityWrite.tsx
@@ -9,6 +9,7 @@ import WritingTitle from './components/writingTitle/WritingTitle';
 import useCommunityWrite from './hooks/UseCommunityWrite';
 import { createPostFormData } from './utils/FormDataUtils';
 import { postBoard } from '@apis/board';
+import { MYPAGE_QUERY_KEY } from '@apis/user';
 import ImgPopupl84 from '@assets/svgs/ImgPopupLogout84';
 import ToolListBanner from '@components/banner/ToolListBanner';
 import CircleButton from '@components/button/circleButton/CircleButton';
@@ -16,7 +17,6 @@ import { AlterModal } from '@components/modal';
 import Title from '@components/title/Title';
 import Toast from '@components/toast/Toast';
 import { useModal } from '@pages/community/hooks';
-import { MYPAGE_QUERY_KEY } from '@pages/myPage/apis/queries';
 
 const CommunityWrite = () => {
   const {
