--- src/pages/signUp/types/index.ts
@@ -1,24 +1,0 @@
-export interface SignupRequest {
-  nickname: string;
-  positions: string;
-  email: string;
-}
-
-export interface SignupResponse {
-  statusCode: number;
-  message: string;
-  data: {
-    nickname: string;
-    positions: string;
-    email: string;
-    jwtTokenResponse: {
-      accessToken: string;
-      refreshToken: string;
-    };
-  };
-}
-
-export interface ErrorResponse {
-  status: number;
-  message: string;
-}
